+++ tests/test_processor.py
@@ -190,7 +190,7 @@ import shutil
 import inspect
 from pyspider.database.sqlite import projectdb
 from pyspider.processor.processor import Processor
-from pyspider.libs.base_queue import get_queue as Queue
+from pyspider.libs.queue import get_queue as Queue
 from pyspider.libs.utils import run_in_thread
 from pyspider.libs import sample_handler
